i2eChatbot/requirements.txt
python-dotenv
faiss-cpu
langchain-huggingface
langchain-groq
langchain
streamlit
fastapi
uvicorn
langchain-community